drops old womenwriters index code. adds women_writers_index.rb.
