+++ timbuctoo-instancev4/src/main/scripts/index_scripts/womenwriters/mappers/ww_person_mapper.rb
@@ -1,5 +1,4 @@
 require File.dirname(__FILE__) + '/../../lib/timbuctoo_solr/default_mapper'
-
 require File.dirname(__FILE__) + '/../../lib/mixins/converters/to_year_converter'
 require File.dirname(__FILE__) + '/../../lib/mixins/converters/to_names_converter'
