--- timbuctoo-instancev4/src/main/scripts/index_scripts/womenwriters_v1/.gitignore
@@ -1,1 +1,0 @@
-*.json
